requirements.txt
paho-mqtt>=1.3.0
netifaces>=0.10.6